+++ HISTORY.md
@@ -13,6 +13,7 @@
 
 ### 2025-12-02
 * widgets, gui: prevent a launched tutorial stage from automatically getting keyboard focus
+* objects, gui: disable drag & drop for assets on a launched tutorial stage
 
 ### 2025-10-31
 * store, scenes, gui: auto-launch tutorials in template projects, don't save tutorials in template copies